move functions around, remove to_pdf entry point in anticipation of cleanup, add lint stage
