+++ .github/workflows/pipeline.yml
@@ -13,7 +13,27 @@ on:
   workflow_dispatch:
 
 jobs:
+
+  lint:
+    name: Lint with Black
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: '3.11'
+
+      - name: Install Black
+        run: pip install black
+
+      - name: Run Black in check mode
+        run: black --check .
+
   build:
+    needs: lint
     runs-on: windows-latest
 
     steps:
