+++ source/autologic.py
@@ -210,10 +210,6 @@ def main(
     print(f"\n  ---\n\n  >>> Iteration {iteration} accepted <<<")
 
     # export data
-    work_assignments = event.get_work_assignments()
-    heat_assignments = event.get_heat_assignments()
-    utils.autologic_event_to_csv(work_assignments)
-    utils.autologic_event_to_pdf(
-        work_assignments=work_assignments, heat_assignments=heat_assignments
-    )
+    event.to_csv()
+    event.to_pdf()
     print()
